src/main/java/org/jenkinsci/plugins/deploydb/DeployDbConfig.java
package org.jenkinsci.plugins.deploydb;

import hudson.Extension;
import jenkins.model.GlobalConfiguration;
import net.sf.json.JSONObject;
import org.kohsuke.stapler.StaplerRequest;

@Extension
public class DeployDbConfig extends GlobalConfiguration {

    private String baseUrl;

    public DeployDbConfig() {
        load();
    }

    public String getBaseUrl() {
        return baseUrl;
    }

    public void setBaseUrl(String baseUrl) {
        this.baseUrl = baseUrl;
    }

    @Override
    public boolean configure(StaplerRequest req, JSONObject json) throws FormException {
        req.bindJSON(this, json);
        save();
        return true;
    }

}
